Rename UI elements from the About Dialog class
The UI elements for the AboutDialog class were renamed in order to allow easier identification of the desired elements in the source code.
I also added layout elements to this form so the window wouldn't be deformed when resized.
Beyond that, I also added a few aesthetic enhancements to logos used in the project.