Remove unused excession handling from the GUI
The CustonException exception handling in the GUI was trying to catch an exception that isn't thown in any part of the GUI code, the only function it was serving was cluttering the code.
Besides, the Qt framework and C++ standards do not recommend the use of exceptions unless the code has a lot of tests and special care is taken in order to avoid dangling pointer and other issues related to the breaking of the code control flow.