Fix main window crashing after closing modelsetupdialog
After closing the model setup dialog, there was an error within the main window, causing the application to crash.
The hardware in the loop configurations was causing this error because it used to set its options relying on the order of the items in the tree widget.
To fix this, the method loops through the items, check its name, and change the appropriate ones.