Add a ProcessOutputWindow to run gazebo
I have replaced the std::system calls for the roslaunch file used to open the simulation with a call to a QProcess.

The GUI nows shows a windows wich reports all errors and the full output of the simulation process.

OBS: I haven't done any modifications to the hardware in the loop sections. This needs to be refactored in a later time.