Add new version of the launch files
The launch files and the GUI were modified in order to allow the specification of the desired world for use in the simulation as a parameter of the ROS launch file.

Previously, a new launch file was generated with the world path hardcoded into it.

The GUI was also modified so that new worlds have an adequate value in the name tag.

Finally, the existing worlds were updated in order to remove specific paths from their description files.