Fix std::logic_error when reading environment variables
When running the controller node the program was crashing with the following message:

terminate called after throwing an instance of 'std::logic_error'
  what():  basic_string::_M_construct null not valid
Abortado (imagem do núcleo gravada)

This was caused by an issue of the std::string class when receiving a null pointer in its constructor, what was happening when the program read the environment variables.

This commit fix this issues and adds a few debug messages used to diagnose this problem.