Add methods to kill child process to ProcessOutputWindow
Added a method to the ProcessOutputWindow to make sure that the entire process tree is finished in the case of a forced termination.

Without this, closing the process window didn't kill the child process, and this was causing the gazebo server to always remain open, what causes problems when starting a new simulation instance.