Add implementation of SGR ANSI Escape Codes
Add an implementation that allows programs that are executed in the ProcessOutputWindow to set paramters configuring how the text is printed, such as color, background color, font family, font weight, underline, overline, etc.
All of the 107 SGR escape codes are implemented, with the exception of Blink that could not be implemented in any reasonable way.
Framed and encircled text are implemented in particular ways, all of the other codes follow the standard implementations, including 88 colors compatibility.