Fix application crash while trying to open model setup dialog
If an user tried to open the Model Setup dialog for a model that did not exist, or that doesn't have a config.xml and a correctly named SDF file, the application crashed with the menssage "Problemas no arquivo XML".

The correction is based on a check of the necessary files existance, warning dialogs informing the problem and a possible solution and log messages in the correct points.