Deprecate TreeItens class
The TreeItens class was merely a namespace for a few static functions, that forced unnecessary conversions between QString and std::strings to allow its use.

These methods are deprecated and will be removed in the next version.
The new interface is based on functions without a classs and with a standardized usage of QStrings.