Fix path of the log files for the DataSaveTiltRotor plugin
The simulation of the UAV 4 was creating a series of simulation log files inside the source/Structure folder.

These files should be created inside the source/structure/Matlab folder.

The error was caused because there was a missing forward slash in the code that initializes these files, so for example:
The /source/Structure/Matlab/DAl.txt file was being saved as /source/Structure/MatlabDAl.txt. (note the missing forward slash after Matlab).

To correct the error, the new code simply checks if the string is terminated by a forward slash, and adds one if it is missing.